+++ HISTORY.md
@@ -1,5 +1,8 @@
 History of features added to LiChess Tools in time (reversed order)
 
+21 Aug
+  - added computer eval for not logged in players and enabled auto annotations for them
+
 09 Aug
   - ceval line options - highlight